+++ src/templates/mediaproject/src/components/withError/index.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+import ErrorBoundary from './ErrorBoundary';
+
+export const withError =
+  (PageComponent: React.FC<React.ComponentProps<any>>) =>
+  (props: React.ComponentProps<any>) => {
+    return (
+      <ErrorBoundary>
+        <PageComponent {...props} />
+      </ErrorBoundary>
+    );
+  };
+
+export default withError;