+++ src/templates/mediaproject/src/store/RootStore.ts
@@ -1,0 +1,7 @@
+import { BaseRootStore } from '@ktsstudio/mediaproject-stores';
+
+import { UserStore } from './UserStore';
+
+export class RootStore extends BaseRootStore {
+  userStore = new UserStore(this);
+}