+++ src/templates/mediaproject/src/types/window.d.ts
@@ -1,0 +1,6 @@
+import { WindowType } from '@ktsstudio/mediaproject-utils';
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  interface Window extends WindowType {}
+}