+++ src/templates/vk-mini-app/src/types/urls.ts
@@ -1,0 +1,6 @@
+import { Method } from 'axios';
+
+export type UrlConfigType = {
+  url: string;
+  method?: Method;
+};